Relocate healthyController files
