+++ src/server/app.ts
@@ -3,7 +3,7 @@ import morgan from "morgan";
 import { generalError } from "./error/generalError.js";
 import monumentsRouter from "../monument/router/monumentsRouter.js";
 import cors from "cors";
-import HealthyCheckController from "../HealthyCheckController/HealthyCheckController.js";
+import HealthyCheckController from "./HealthyCheckController/HealthyCheckController.js";
 import serverRouter from "./router/serverRouter.js";
 
 const app = express();
